+++ Changelog.md
@@ -2,6 +2,9 @@
 
 [full changelog](https://github.com/Mange/roadie/compare/v1.0.2...master)
 
+* Bug fixes
+  * Don't change `asset_providers` of a `Roadie::Document` when applying Options with no `asset_providers` set.
+
 ### 1.0.2
 
 [full changelog](https://github.com/Mange/roadie/compare/v1.0.1...v1.0.2)